Refactor : Move StageModel to dto package
