+++ src/main/java/hatch/hatchserver2023/domain/stage/dto/StageResponseDto.java
@@ -1,7 +1,6 @@
 package hatch.hatchserver2023.domain.stage.dto;
 
 import com.fasterxml.jackson.annotation.JsonInclude;
-import hatch.hatchserver2023.domain.stage.StageModel;
 import hatch.hatchserver2023.domain.talk.domain.TalkMessage;
 import hatch.hatchserver2023.domain.talk.dto.TalkResponseDto;
 import lombok.*;
